Create Enlightenment
Pepsiman[a] is an action video game developed and published by KID for the PlayStation. It was released in Japan on March 4, 1999, and is based on the eponymous Japanese superhero mascot for the American carbonated soft drink Pepsi. It focuses the player on avoiding obstacles by running, dashing, and jumping, while Pepsiman automatically runs forward through each of the game's stages.

The game was made on a low budget, prompting the decision to make videos in-between stages that show a man drinking Pepsi, as they were cheap to produce. The game also features 3D cutscenes, for which the future visual novel writer Kotaro Uchikoshi created 3D models. While an American publisher did look into acquiring the rights to publish the game in the United States, it remained a Japan-exclusive game.

Reviewers frequently compared Pepsiman to other games, including Crash Bandicoot, and commented on its simplicity and its price, which was thought to be low. A writer for Complex included it on a list of company-branded games that "didn't suck", commenting that it is not a bad game as long as the player can tolerate the large amount of advertisement in it. According to Uchikoshi, the game did not sell well.

Contents
1	Gameplay
2	Background and development
3	Reception
4	Notes
5	References
6	External links
Gameplay[edit]

Gameplay of Pepsiman. Here, he is jumping on an intersection in an attempt to avoid a moving car.
Pepsiman is an action game[2] that consists of four stages, each divided into smaller segments,[3] and each involving the superhero Pepsiman saving a person who is dehydrated, such as a military man in the middle of a desert, by giving him a can of Pepsi.[4] The first three stages are based on real locations, San Francisco, New York City and Texas. The last one takes place in Pepsi City.[5] The game is played from a third-person perspective, with Pepsiman automatically running forward through the stages,[4] sometimes running through homes and other buildings.[3] The player takes control of Pepsiman himself, aiming to dodge obstacles, such as cars, construction cranes, and people,[4] as well as Pepsi-branded obstacles, including a Pepsi truck.[6] The player does this by using four different moves: running, dashing, jumping, and super-jumping.[3] The player gains points by collecting Pepsi cans.[6]

In some stages, Pepsiman's head becomes stuck inside a steel drum, which inverts the controls, and in some, he rides on a skateboard, which requires to player to avoid all obstacles. Throughout each stage is a number of checkpoints; if Pepsiman gets hit by obstacles too many times, the player is required to restart from the latest checkpoint. Each stage ends with Pepsiman being chased by an object,[4] such as a giant Pepsi can.[6] In between stages, the player is shown videos of an American man (played by Mike Butters) drinking Pepsi and eating chips and pizza as he watches Pepsiman.[4]

Background and development[edit]
Pepsiman is based on Pepsi's mascot of the same name, which was created for Pepsi's Japanese branch.[4] The character, whose fictional backstory says he used to be a scientist who transformed into a superhero after coming into contact with "Holy Pepsi",[7] was featured in Japanese Pepsi commercials[4] and in the Japanese version of the video game Fighting Vipers; he became popular in Japan,[8] spawning related characters such as Lemon Pepsiman and Pepsiwoman,[7] and Pepsi decided to promote the character with a video game.[4]

The game was developed by the Japanese video game developer KID. It was made on a low budget, which led to the decision to make the low-cost video scenes of actor Mike Butters drinking Pepsi.[9][10] The game also uses 3D event scenes, which were modeled by Kotaro Uchikoshi, who would later be a scenario writer for visual novels at KID. This was Uchikoshi's first job; he had been hired to plan video game adaptations of board games, but ended up being part of the development of Pepsiman instead, which was already in progress when he joined KID in 1998.[9] The game was released in Japan by KID for the PlayStation on March 4, 1999;[2] while an American publisher was looking into acquiring the rights to publish the game in the United States,[8] it remained Japan-exclusive. Despite this, the game is entirely in English, not Japanese (although with Japanese subtitles for dialogue).[4] According to Uchikoshi, the game did not sell well.[9]

In 2019, the game was featured in an episode of James Rolfe's comedy web series Angry Video Game Nerd, in which Butters reprised his role from the game's cutscenes.[11] The soundtrack for the game received a vinyl release in 2020 by the European label Chipped Records.[12]

Reception[edit]
Reception
Review scores
Publication	Score
Famitsu	25/40[2]
GameFan	90/100[13]
Joypad	2/10[14]
Gamers' Republic	B−[15]
Planet Playstation	88/100[16]
Writers for Famitsu called the game "super-simple", comparing it to Metro-Cross and Paperboy, and calling it a simplified version of Crash Bandicoot.[2] Others have made similar comments. A reviewer for IGN also compared it to Crash Bandicoot, described the gameplay as "simplistic [and] route memorization-based", and said that the thing the game would be remembered for was its "extremely bizarre premise". They still felt that the game was not bad, and that it was worth the price, which they noted was low.[3] James Mielke at GameSpot called the game a "nifty little distraction", and said that the gameplay was similar to the "old-school gaming dynamics of yesteryear". He commented on the low price, but said that it was difficult to find imports of it.[8] Gamers' Republic magazine rated the game a B−.[15] Gamers' Republic later listed the game in their 1999 Video Game Buyers Guide and Y2K Preview as one of the best games to import from Japan that year.[17]

In 2011, Allistair Pinsof at Destructoid reviewed the game, calling it a mix between Paperboy and Muscle March in terms of the complexity and pace, and compared the gameplay to Crash Bandicoot. He found it to be "such a gloriously twisted, charming spectacle" that it would be difficult not to like it; he said that the main reason to play the game is "the sheer lunacy" of it, saying that the game is "obsessed" with America, and portrays Americans as "unhygienic hillbillies" in a manner that makes it unclear if it is a self-aware parody or not. He concluded that the game is funny, but not great, and that the ridiculous premise and its large amount of small details make the game "charmingly brain-dead".[4] In 2013, Justin Amirkhani at Complex included the game in a list of company-branded video games that "didn't suck", saying that while the game's graphics had not aged well, it was mechanically very similar to Temple Run, which Amirkhani called his favorite iOS game. He concluded that Pepsiman is not a bad game for people with quick reflexes, as long as they can stand the high amount of advertising within the game; he claimed that Pepsiman was the advergame with the largest amount of "logos-per-second".[6]

In 2015, Retro Gamer magazine listed it as number 18 on their list of "The 20 Greatest PlayStation Games You've Never Played".[18